requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.1.8
Django==4.2.24
django-environ==0.12.0
django-tinymce==4.1.0
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
feedparser==6.0.12
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sgmllib3k==1.0.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.10.0
